+++ pyemma/coordinates/tests/test_datainmemory.py
@@ -8,11 +8,44 @@ import unittest
 from pyemma.coordinates.io.data_in_memory import DataInMemory
 from pyemma.util.log import getLogger
 import numpy as np
+from pyemma.coordinates.api import kmeans
+
+import tempfile
+import os
 
 logger = getLogger('TestDataInMemory')
 
 
 class TestDataInMemory(unittest.TestCase):
+
+    @classmethod
+    def setUpClass(cls):
+        d = np.random.random((100, 3))
+        d_1d = np.random.random(100)
+
+        f1 = tempfile.mktemp()
+        f2 = tempfile.mktemp(suffix='.npy')
+        f3 = tempfile.mktemp()
+        f4 = tempfile.mktemp(suffix='.npy')
+
+        np.savetxt(f1, d)
+        np.save(f2, d)
+
+        np.savetxt(f3, d_1d)
+        np.save(f4, d_1d)
+
+        cls.files2d = [f1, f2]
+        cls.files1d = [f3, f4]
+        cls.d = d
+        cls.d_1d = d_1d
+        return cls
+
+    @classmethod
+    def tearDownClass(cls):
+        for f in cls.files1d:
+            os.remove(f)
+        for f in cls.files2d:
+            os.remove(f)
 
     def testListOfArrays(self):
 
@@ -37,69 +70,48 @@ class TestDataInMemory(unittest.TestCase):
         self.assertEqual(
             d.trajectory_lengths(), [frames_per_traj for i in xrange(1)])
 
-    def testChunkedAccess(self):
-        frames_per_traj = 100
-        dim = 3
-        data = [np.random.random((frames_per_traj, dim)) for i in xrange(3)]
+    def test1dData(self):
+        n = 3
+        data = np.arange(n)
+        reader = DataInMemory(data)
 
-        d = DataInMemory(data)
-        d.chunksize = 6
-        itraj = 0
-        last_chunk = False
-        while not last_chunk:
-            last_chunk_in_traj = False
-            t = 0
-            while not last_chunk_in_traj:
-                # iterate over times within trajectory
-                X = d.next_chunk()
-                L = np.shape(X)[0]
-                # last chunk in traj?
-                last_chunk_in_traj = (t + 0 + L >= d.trajectory_length(itraj))
-                # last chunk?
-                last_chunk = (
-                    last_chunk_in_traj and itraj >= d.number_of_trajectories() - 1)
-                # increment time
-                t += L
-            # increment trajectory
-            itraj += 1
+        self.assertEqual(reader.trajectory_lengths(), [n])
+        self.assertEqual(reader.dimension(), 3)
+        self.assertEqual(reader.ntraj, 1)
+        self.assertEqual(reader.n_frames_total(), 3)
 
-    @unittest.skip("known to be broken.")
-    def testChunkedLaggedAccess(self):
-        frames_per_traj = 100
-        dim = 3
-        lag = 3
-        chunksize = 6
-        data = [np.random.random((frames_per_traj, dim)) for i in xrange(3)]
+        k = kmeans(data, k=2)
+        print k.dtrajs
 
-        d = DataInMemory(data)
-        d.chunksize = chunksize
-        itraj = 0
-        last_chunk = False
-        ipass = 3
-        while ipass > 0:
-            logger.debug("ipass : %i " % ipass)
-            while not last_chunk:
-                last_chunk_in_traj = False
-                t = 0
-                while not last_chunk_in_traj:
-                    # iterate over times within trajectory
-                    X, Y = d.next_chunk(lag=lag)
-                    assert X is not None
-                    assert Y is not None
-                    L = np.shape(X)[0]
-                    K = np.shape(Y)[0]
-                    assert L == K, "data in memory gave different chunksizes"
-                    # last chunk in traj?
-                    last_chunk_in_traj = (
-                        t + lag + L >= d.trajectory_length(itraj))
-                    # last chunk?
-                    last_chunk = (
-                        last_chunk_in_traj and itraj >= d.number_of_trajectories() - 1)
-                    # increment time
-                    t += L
-                # increment trajectory
-                itraj += 1
-            ipass -= 1
+    def test1dDataList_diff_dim(self):
+        n = 3
+        data = [np.arange(n), np.arange(n + 1)]
+        with self.assertRaises(ValueError):
+            reader = DataInMemory(data)
+
+    def test1dDataList(self):
+        n = 10
+        data = [np.arange(n), np.arange(n)]
+        reader = DataInMemory(data)
+
+        self.assertEqual(reader.trajectory_lengths(), [1, 1])
+        self.assertEqual(reader.dimension(), n)
+        self.assertEqual(reader.ntraj, 2)
+        self.assertEqual(reader.n_frames_total(), 2)
+
+    def test_file_1d(self):
+        DataInMemory(self.files1d)
+
+    def test_file_2d(self):
+        DataInMemory(self.files2d)
+
+    def testNotEqualDims(self):
+        """ should raise, since different dims can not be processed"""
+        data = [np.zeros((10, 3)), np.zeros((10, 5))]
+
+        with self.assertRaises(ValueError):
+            DataInMemory(data)
+
 
 if __name__ == "__main__":
     unittest.main()