+++ pyemma/msm/ui/timescales.py
@@ -8,14 +8,17 @@ __docformat__ = "restructuredtext en"
 __all__=['ImpliedTimescales']
 
 import numpy as np
+import warnings
 
-from pyemma.msm.estimation import cmatrix, connected_cmatrix, tmatrix, bootstrap_counts
+from pyemma.msm.estimation import cmatrix, number_of_states, connected_cmatrix, tmatrix, bootstrap_counts
 from pyemma.msm.analysis import timescales
 from pyemma.util.statistics import confidence_interval
+from pyemma.util.types import ensure_dtraj_list as _ensure_dtraj_list
 
-#TODO: connectivity is currently not used. Introduce different connectivity modes (lag, minimal, set)
+#TODO: connectivity flag is currently not used. Introduce different connectivity modes (lag, minimal, set)
 #TODO: if not connected, might add infinity timescales.
 #TODO: Timescales should be assigned by similar eigenvectors rather than by order
+#TODO: when requesting too long lagtimes, throw a warning and exclude lagtime from calculation, but compute the rest
 class ImpliedTimescales(object):
     
     # estimated its. 2D-array with indexing: lagtime, its
@@ -24,7 +27,7 @@ class ImpliedTimescales(object):
     _its_samples = None
     
     
-    def __init__(self, dtrajs, lags = None, nits = 10, connected = True, reversible = True):
+    def __init__(self, dtrajs, lags = None, nits = 10, connected = True, reversible = True, failfast = False):
         r"""Calculates the implied timescales for a series of lag times.
         
         Parameters
@@ -40,36 +43,45 @@ class ImpliedTimescales(object):
             compute the connected set before transition matrix estimation at each lag
             separately
         reversible = True : boolean
-            estimate the transition matrix reversibly (True) or nonreversibly (False)"""
+            estimate the transition matrix reversibly (True) or nonreversibly (False)
+        failfast = False : boolean
+            if True, will raise an error as soon as not all requested timescales can be computed at all requested
+            lagtimes. If False, will continue with a warning and compute the timescales/lagtimes that are possible.
+
+        """
         # initialize
-        self._dtrajs = dtrajs
+        self._dtrajs = _ensure_dtraj_list(dtrajs)
         self._connected = connected
         self._reversible = reversible
-        # maximum trajectory length and number of states
 
-        # Check if input was array or list of arrays
-        if np.ndim(dtrajs) < 2:
-           # We we were only parsed an array
-           dtrajs=[dtrajs]
-        lengths = np.zeros(len(dtrajs))
-
-        n = 0
-        for i in range(len(dtrajs)):
-            lengths[i] = len(dtrajs[i])
-            n = max(n, np.max(dtrajs[i]))
         # maximum number of timescales
-        self._nits = min(nits, n)
+        nstates = number_of_states(dtrajs)
+        self._nits = min(nits, nstates-1)
+
+        # trajectory lengths
+        self.lengths = np.zeros(len(dtrajs))
+        for i in range(len(dtrajs)):
+            self.lengths[i] = len(dtrajs[i])
+        self.maxlength = np.max(self.lengths)
+
         # lag time
         if (lags is None):
-            maxlag = 0.5 * np.sum(lengths) / float(len(lengths))
-            self._lags = self.__generate_lags__(maxlag, 1.5)
+            maxlag = 0.5 * np.sum(self.lengths) / float(len(self.lengths))
+            self._lags = self._generate_lags(maxlag, 1.5)
         else:
-            self._lags = lags
+            self._lags = np.array(lags)
+            # check if some lag times are forbidden.
+            if np.max(self._lags) >= self.maxlength:
+                Ifit = np.where(self._lags < self.maxlength)[0]
+                Inofit = np.where(self._lags >= self.maxlength)[0]
+                warnings.warn('Some lag times exceed the longest trajectories. Will ignore lag times: '+str(self._lags[Inofit]))
+                self._lags = self._lags[Ifit]
+
         # estimate
-        self.__estimate__()
+        self._estimate()
 
                 
-    def __generate_lags__(self, maxlag, multiplier):
+    def _generate_lags(self, maxlag, multiplier):
         r"""Generate a set of lag times starting from 1 to maxlag, 
         using the given multiplier between successive lags
         
@@ -85,7 +97,7 @@ class ImpliedTimescales(object):
         return lags
 
 
-    def __C_to_ts__(self, C, tau):
+    def _estimate_ts_tau(self, C, tau):
         r"""Estimate timescales from the given count matrix.
         
         """
@@ -101,26 +113,41 @@ class ImpliedTimescales(object):
             return None # no timescales available
         
     
-    def __estimate__(self):
+    def _estimate(self):
         r"""Estimates ITS at set of lagtimes
         
         """
         # initialize
         self._its = np.zeros((len(self._lags), self._nits))
+        maxnits = self._nits
+        maxnlags  = len(self._lags)
         for i in range(len(self._lags)):
+            # get lag time to be used
             tau = self._lags[i]
-            # estimate count matrix
+            # unconnected C matrix
             C = cmatrix(self._dtrajs, tau)
             # estimate timescales
-            ts = self.__C_to_ts__(C, tau)
+            ts = self._estimate_ts_tau(C, tau)
             if (ts is None):
-                raise RuntimeError('Could not compute a single timescale at tau = '+str(tau)+
-                                   '. Probably a connectivity problem. Try using smaller lagtimes')
-            if (len(ts) < self._nits):
-                raise RuntimeError('Could only compute '+str(len(ts))+' timescales at tau = '+str(tau)+
-                                   ' instead of the requested '+str(self._nits)+'. Probably a '+
-                                   ' connectivity problem. Request less timescales or smaller lagtimes')
+                maxnlags = i
+                warnings.warn('Could not compute a single timescale at tau = '+str(tau)+
+                              '. Probably a connectivity problem. Try using smaller lagtimes')
+                break
+            elif (len(ts) < self._nits):
+                maxnits = min(maxnits, len(ts))
+                warnings.warn('Could only compute '+str(len(ts))+' timescales at tau = '+str(tau)+
+                              ' instead of the requested '+str(self._nits)+'. Probably a '+
+                              ' connectivity problem. Request less timescales or smaller lagtimes')
             self._its[i,:] = ts
+
+        # any infinities?
+        if (np.any(np.isinf(self._its))):
+            warnings.warn('Timescales contain infinities, indicating that the data is disconnected at some lag time')
+
+        # clean up
+        self._nits = maxnits
+        self._lags = self._lags[:maxnlags]
+        self._its = self._its[:maxnlags][:,:maxnits]
 
     
     def bootstrap(self, nsample=10):
@@ -129,42 +156,75 @@ class ImpliedTimescales(object):
         """
         # initialize
         self._its_samples = np.zeros((len(self._lags), self._nits, nsample))
+        self._nits_sample = self._nits
+        maxnits = self._nits_sample
+        maxnlags  = len(self._lags)
         for i in range(len(self._lags)):
             tau = self._lags[i]
-            sampledWell = True
+            all_ts = True
+            any_ts = True
             for k in range(nsample):
                 # sample count matrix
                 C = bootstrap_counts(self._dtrajs, tau)
                 # estimate timescales
-                ts = self.__C_to_ts__(C, tau)
+                ts = self._estimate_ts_tau(C, tau)
                 # only use ts if we get all requested timescales
-                if (ts != None):
+                if (ts is not None):
                     if (len(ts) == self._nits):
                         self._its_samples[i,:,k] = ts
                     else:
-                        sampledWell = False
+                        all_ts = False
+                        maxnits = min(maxnits, len(ts))
+                        self._its_samples[i,:maxnits,k] = ts
                 else:
-                    sampledWell = False
-            if (not sampledWell):
-                raise RuntimeWarning('Could not compute all requested timescales at tau = '+str(tau)+
+                    any_ts = False
+                    maxnlags = i
+            if (not all_ts):
+                warnings.warn('Could not compute all requested timescales at tau = '+str(tau)+
                                      '. Bootstrap is incomplete and might be non-representative.'+
                                      ' Request less timescales or smaller lagtimes')
+            if (not any_ts):
+                warnings.warn('Could not compute a single timescale at tau = '+str(tau)+
+                              '. Probably a connectivity problem. Try using smaller lagtimes')
+        # clean up
+        self._nits_sample = maxnits
+        self._lags_sample = self._lags[:maxnlags]
+        self._its_samples = self._its_samples[:maxnlags,:,:][:,:maxnits,:]
 
-    
+
+    @property
+    def lagtimes(self):
+        r"""Return the list of lag times for which timescales were computed.
+
+        """
+        return self._lags
 
     def get_lagtimes(self):
         r"""Return the list of lag times for which timescales were computed.
         
         """
+        warnings.warn('get_lagtimes() is deprecated. Use lagtimes', DeprecationWarning)
         return self._lags
 
-
+    @property
     def number_of_timescales(self):
         r"""Return the number of timescales.
         
         """
         return self._nits
 
+    @property
+    def timescales(self):
+        r"""Returns the implied timescale estimates
+
+        Returns
+        --------
+        timescales : ndarray((l x k), dtype=float)
+            timescales for all processes and lag times.
+            l is the number of lag times and k is the number of computed timescales.
+
+        """
+        return self.get_timescales()
 
     def get_timescales(self, process = None):
         r"""Returns the implied timescale estimates
@@ -188,14 +248,46 @@ class ImpliedTimescales(object):
         else:
             return self._its[:,process]
 
+
+    @property
     def samples_available(self):
         r"""Returns True if samples are available and thus sample
         means, standard errors and confidence intervals can be
         obtained
         
         """
-        return (self._its_samples != None)
+        return (self._its_samples is not None)
 
+
+    @property
+    def sample_lagtimes(self):
+        r"""Return the list of lag times for which sample data is available
+
+        """
+        return self._lags_sample
+
+
+    @property
+    def sample_number_of_timescales(self):
+        r"""Return the number of timescales for which sample data is available
+
+        """
+        return self._nits_sample
+
+
+    @property
+    def sample_mean(self):
+        r"""Returns the sample means of implied timescales. Need to
+        generate the samples first, e.g. by calling bootstrap
+
+        Returns
+        -------
+        timescales : ndarray((l x k), dtype=float)
+            mean timescales for all processes and lag times.
+            l is the number of lag times and k is the number of computed timescales.
+
+        """
+        return self.get_sample_mean()
 
     def get_sample_mean(self, process = None):
         r"""Returns the sample means of implied timescales. Need to
@@ -225,8 +317,25 @@ class ImpliedTimescales(object):
             return np.mean(self._its_samples[:,process,:], axis = 1)
 
 
+    @property
+    def sample_std(self):
+        r"""Returns the standard error of implied timescales. Need to
+        generate the samples first, e.g. by calling bootstrap
+
+
+        Returns
+        -------
+        Returns
+        -------
+        timescales : ndarray((l x k), dtype=float)
+            standard deviations of timescales for all processes and lag times.
+            l is the number of lag times and k is the number of computed timescales.
+
+        """
+        return self.get_sample_std()
+
     def get_sample_std(self, process = None):
-        r"""Returns the sample means of implied timescales. Need to
+        r"""Returns the standard error of implied timescales. Need to
         generate the samples first, e.g. by calling bootstrap
         
         Parameters