+++ tools/ci/travis/make_docs.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 function install_deps {
-	sudo apt-get install -qq pandoc # texlive-latex-base
+	wget https://github.com/jgm/pandoc/releases/download/1.13.2/pandoc-1.13.2-1-amd64.deb \
+		-O pandoc.deb
+	sudo dpkg -i pandoc.deb
+	#sudo apt-get install -qq pandoc # texlive-latex-base
 	conda install -q --yes $doc_deps
 	pip install -r requirements-build-doc.txt wheel
 }