+++ doc/CHANGELOG.rst
@@ -2,4 +2,4 @@ v 1.1.2 (planned)
 -------
 
 - PCCA++ now produces correct memberships (fixes a problem from nonorthonormal eigenvectors)
-
+- Improved Coordinates API documentation (Examples, examples, EXAMPLES)